# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Economizza Api",
        "description": "Documentação da API",
        "contact": {
            "email": "goiswilliam194@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Lista todos os usuários",
                "operationId": "3ab81e3f9b24b4f36ab08d72349ac632",
                "responses": {
                    "200": {
                        "description": "Usuários listados com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Cria um novo usuário",
                "operationId": "ba548b97148bb5c0d8cab3bc3f128bfd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso"
                    },
                    "400": {
                        "description": "Erro na solicitação"
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Busca um usuário pelo ID",
                "operationId": "daaf7a6169ee9348c3fddb07778598ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Atualiza um usuário",
                "operationId": "59e98fba99e974815e27979dd58e2175",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        },
        "/api/v1/users/social/upsert": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Cria ou atualiza um usuário com base em dados de autenticação social",
                "operationId": "07640e8ee621186b2833987fa986217c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "user": {
                                                "properties": {
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "id": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "session": {
                                                "properties": {
                                                    "access_token": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário criado ou atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Efetua login de um usuário",
                "operationId": "8d0c84b850f0cc589b9aba0f65fe89c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "Estrutura de dados para um usuário",
                "properties": {
                    "id": {
                        "description": "Identificador único do usuário",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Nome do usuário",
                        "type": "string"
                    },
                    "email": {
                        "description": "Endereço de email do usuário",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
...
}